+++ CHANGELOG.md
@@ -5,7 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.2] - [Unreleased]
+## [1.1.3] - [Unreleased]
+
+### Added
+
+- Added an accounts receivable module with automatic loading, local client search, pull-to-refresh
+- Added client account cards showing credit, pending balance, available credit, currency, and credit status.
+- Added document detail modal with invoice and credit note labels, formatted totals, issue dates, observations, and delivery note references.
+
+### Fixed
+
+- Font size viiew in create orders
+
+## [1.1.2] - [2026-06-14]
 
 ### Added